+++ src/router/index.js
@@ -63,20 +63,6 @@ const moderationLog = {
   ]
 }
 
-const emojiPacksDisabled = disabledFeatures.includes('emoji-packs')
-const emojiPacks = {
-  path: '/emoji-packs',
-  component: Layout,
-  children: [
-    {
-      path: 'index',
-      component: () => import('@/views/emoji-packs/index'),
-      name: 'Emoji packs',
-      meta: { title: 'emoji-packs', icon: 'settings', noCache: true }
-    }
-  ]
-}
-
 export const constantRouterMap = [
   {
     path: '/redirect',
@@ -144,7 +130,6 @@ export const asyncRouterMap = [
   ...(invitesDisabled ? [] : [invites]),
   ...(moderationLogDisabled ? [] : [moderationLog]),
   ...(settingsDisabled ? [] : [settings]),
-  ...(emojiPacksDisabled ? [] : [emojiPacks]),
   {
     path: '/users/:id',
     component: Layout,